moved context folder one level up
