+++ frontend/src/components/Items.js
@@ -4,7 +4,7 @@ import { Grid } from "@material-ui/core";
 
 import Item from "./Item";
 import { hasVotedFor } from "../utils/userUtils";
-import { BoardContext } from "./context/BoardContext";
+import { BoardContext } from "../context/BoardContext";
 
 function Items(props) {
   const { items, openSnackbar } = props;
